+++ AIServerCore/src/controller/ChatHandler.cpp
@@ -1,5 +1,6 @@
 
 #include "controller/ChatHandler.h"
+#include "Common/Http/ApiResult.h"
 
 void ChatHandler::handle(const http::HttpRequest& req, http::HttpResponse* resp)
 {
@@ -9,9 +10,7 @@ void ChatHandler::handle(const http::HttpRequest& req, http::HttpResponse* resp)
         LOG_INFO << "session->getValue(\"isLoggedIn\") = " << session->getValue("isLoggedIn");
         if (session->getValue("isLoggedIn") != "true")
         {
-            json errorResp;
-            errorResp["status"] = "error";
-            errorResp["message"] = "Unauthorized";
+            json errorResp = common::ApiResult::fail(400, "Unauthorized").toJson();
             std::string errorBody = errorResp.dump(4);
 
             server_->packageResp(req.getVersion(), http::HttpResponse::k401Unauthorized, "Unauthorized", true,